Update CHANGELOGs and renamed some files
