+++ packages/web3-eth-abi/test/fixtures/get_encoded_eip712_data.ts
