+++ packages/web3-types/CHANGELOG.md
@@ -141,3 +141,8 @@ Documentation:
 -   type `Filter` includes `blockHash` (#6206)
 
 ## [Unreleased]
+
+### Added
+
+-   `eth_signTypedData` and `eth_signTypedData_v4` to `web3_eth_execution_api` (#6286)
+-   `Eip712TypeDetails` and `Eip712TypedData` to `eth_types` (#6286)
